# src/main/java/whnazv/biblioteca/infrastructure/adapter/in/dto/UserLoginDto.java
package whnazv.biblioteca.infrastructure.adapter.in.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UserLoginDto {
    private String email;
    private String password;
}